requirements.txt
python-dotenv
streamlit
yfinance
boto3